Fix and Annotate
